(wip) colonies spawn on location properly, ants attempt to spawn properly but some issues happen with collision detection

+++ core/src/aaa/main/game/Perlin.java
@@ -13,7 +13,7 @@ public class Perlin {
         // corresponds to a tile without a wall and 1 corresponds
         // to a tile with a wall.
 
-	public static boolean[][][] generateMap(int width, int height, int seed) {
+	public static boolean[][][] generateMap(int width, int height, long seed) {
 		
 		double[][] wall_array;
 		double[][] resource_array;
